requirements.txt
streamlit==1.29.0
python-binance==1.0.19
pandas==2.1.4
numpy==1.24.3
plotly==5.17.0
ta==0.10.2
python-dotenv==1.0.0
requests==2.31.0
websocket-client==1.6.4